requirements.txt
streamlit
pandas
numpy
scipy
plotly-express
Faker
seaborn
openpyxl
statsmodels
matplotlib